Test: add fixtures for stereo test cases
Also this commit introduces a thin wrapper for basic manipulation for
images.
